+++ Common/CMakeLists.txt
@@ -1,6 +1,6 @@
 add_subdirectory(MathUtils)
 add_subdirectory(Field)
-add_subdirectory(Configuration)
+add_subdirectory(SimConfig)
 
 install(
     DIRECTORY maps
