+++ Common/SimConfig/src/SimConfig.cxx
@@ -8,7 +8,7 @@
 // granted to it by virtue of its status as an Intergovernmental Organization
 // or submit itself to any jurisdiction.
 
-#include <Configuration/SimConfig.h>
+#include <SimConfig/SimConfig.h>
 #include <boost/program_options.hpp>
 #include <iostream>
