+++ Common/SimConfig/test/TestConfig.cxx
@@ -8,7 +8,7 @@
 // granted to it by virtue of its status as an Intergovernmental Organization
 // or submit itself to any jurisdiction.
 
-#include <Configuration/SimConfig.h>
+#include <SimConfig/SimConfig.h>
 #include <iostream>
 
 int main(int argc, char* argv[]) {
